+++ components/ToDo/TodoModal.tsx
@@ -1,187 +1,21 @@
 'use client'
-import { useScrollLock } from '@/src/hooks/useScrollLock'
-import useTodoStore from '@/src/store/todoStore'
+
+import { useTodoModal } from '@/src/hooks/useTodoModal'
 import { Memo } from '@/src/types/memo'
-import { Todo, TodoWithMemo } from '@/src/types/todo'
-import { useSession } from 'next-auth/react'
-import { useParams, useRouter } from 'next/navigation'
-import { useCallback, useEffect, useState } from 'react'
-import { toast } from 'sonner'
 import Button from '../Button'
 import LoadingPage from '../Loading'
 import ConnectMemoBox from '../Memo/ConnectMemoBox'
-import { fetchConnectMemo } from '../Memo/memosServer'
 import { formatDate } from './TodayDateFormat'
-import {
-  deleteTodo,
-  fetchTodos,
-  fetchTodoWithMemo,
-  updateTodo,
-} from './todosServer'
 
 const TodoModal = () => {
-  const { todoId } = useParams()
-  const { data: session } = useSession()
-  const router = useRouter()
-  const unlock = useScrollLock()
-  const [loading, setLoading] = useState(true)
-  const [newTask, setNewTask] = useState<string>('')
-  const [newImportant, setNewImportant] = useState<boolean>(false)
-  const [newCompleted, setNewCompleted] = useState<boolean>(false)
-  const [newDate, setNewDate] = useState<string | null>(null)
-  const [newMemoId, setNewMemoId] = useState<string | null>(null)
-  const [newConnectMemoContent, setNewConnectMemoContent] = useState<string>('')
-  const [newConnect, setNewConnect] = useState<boolean>(false)
-  const [editTodo, setEditTodo] = useState<Todo | null>(null)
-  const [showDeleteConfirm, setShowDeleteConfirm] = useState(false)
-
-  const todolist = useTodoStore((state) => state.todolist)
-  const setTodosStore = useTodoStore((state) => state.setTodosStore)
-
-  const [todoMemo, setTodoMemo] = useState<TodoWithMemo | null>(null)
-  const [isMemoNull, setIsMemoNull] = useState<boolean>(false)
-
-  const [connectMemos, setConnectMemos] = useState<Memo[]>([])
-
-  useEffect(() => {
-    const fetchData = async () => {
-      if (todolist.length === 0 && session?.user?.email) {
-        try {
-          const fetchModalTodos = await fetchTodos(session.user.email)
-          setTodosStore(fetchModalTodos)
-        } catch {
-          toast.error('데이터를 불러오지 못했습니다.')
-        }
-      }
-      setLoading(false)
-    }
-    fetchData()
-  }, [session, todolist.length, setTodosStore])
-
-  useEffect(() => {
-    const fetchTodoWithMemoData = async () => {
-      if (!todoId || typeof todoId !== 'string' || !session?.user?.email) return
-      const todoWithMemo = await fetchTodoWithMemo(todoId, session.user.email)
-      setTodoMemo(todoWithMemo)
-    }
-    fetchTodoWithMemoData()
-  }, [todoId, session])
-
-  const todo = todolist.find((todo: Todo) => todo.id === todoId)
-
-  useEffect(() => {
-    if (!loading && !todo) router.replace('/todo')
-  }, [loading, todo, router])
-
-  useEffect(() => {
-    window.scroll(0, 0)
-  }, [])
-
-  useEffect(() => {
-    const fetchConnectMemoData = async () => {
-      if (session && session.user && session.user.email) {
-        try {
-          const memos = await fetchConnectMemo(session.user.email)
-          setConnectMemos(memos)
-        } catch {
-          toast.error('메모 목록을 불러오지 못했습니다.')
-        }
-      }
-    }
-    fetchConnectMemoData()
-  }, [session])
-
-  const NewConnectNull = () => {
-    setNewConnectMemoContent('')
-    setNewMemoId(null)
-    setIsMemoNull(true)
-  }
-
-  const MemoIDContent = (id: string, content: string) => {
-    setNewConnectMemoContent(content)
-    setNewMemoId(id)
-  }
-
-  const onClose = useCallback(() => {
-    unlock()
-    router.back()
-  }, [router, unlock])
-
-  const handleDeleteTodo = async (todoId: string) => {
-    if (!session?.user?.email) return
-    try {
-      await deleteTodo(todoId, session.user.email)
-      setTodosStore(todolist.filter((todo) => todo.id !== todoId))
-    } catch {
-      toast.error('삭제에 실패했습니다.')
-      return
-    }
-    unlock()
-    router.push('/todo')
-  }
-  const handleEditTodo = (todo: Todo) => {
-    if (!todo) return
-    setEditTodo(todo)
-    setNewTask(todo.task)
-    setNewImportant(todo.important)
-    setNewCompleted(todo.completed)
-    setNewDate(todo.date || null)
-    setNewMemoId(todo.memo_id || null)
-  }
-  const updateTodoInput = async () => {
-    if (!editTodo || !session?.user?.email) return
-
-    const updatedDate = newDate === '' ? null : newDate
-    const updatedTodoId = newMemoId === '' ? null : newMemoId
-    const updatedTodo = {
-      ...editTodo,
-      task: newTask,
-      important: newImportant,
-      completed: newCompleted,
-      date: updatedDate,
-      memo_id: updatedTodoId,
-    }
-    try {
-      const updatedTodos = await updateTodo(
-        editTodo.id,
-        updatedTodo,
-        session.user.email
-      )
-
-      setTodosStore((prev) =>
-        prev.map((todo) => {
-          const updated = updatedTodos.find((t) => t.id === todo.id)
-          return updated ? updated : todo
-        })
-      )
-
-      const updatedTodoWithMemo = await fetchTodoWithMemo(
-        editTodo.id,
-        session.user.email
-      )
-      setTodoMemo(updatedTodoWithMemo)
-
-      setEditTodo(null)
-      setNewTask('')
-      setNewImportant(false)
-      setNewCompleted(false)
-      setNewDate(null)
-      setNewMemoId(null)
-      setNewConnect(false)
-      setIsMemoNull(false)
-    } catch {
-      toast.error('수정에 실패했습니다.')
-    }
-  }
-  const handleClearDate = () => {
-    setNewDate(null)
-  }
+  const { state, actions } = useTodoModal()
+  const { todo, loading, editTodo } = state
 
   if (loading) {
     return (
       <div
         className="fixed inset-0 z-[100] bg-black/30 flex justify-center items-center p-2"
-        onClick={onClose}
+        onClick={actions.onClose}
       >
         <div onClick={(e) => e.stopPropagation()}>
           <LoadingPage />
@@ -195,7 +29,7 @@ const TodoModal = () => {
   return (
     <div
       className="fixed inset-0 z-[100] bg-black/30 flex justify-center items-center p-2"
-      onClick={onClose}
+      onClick={actions.onClose}
     >
       <div
         className="relative w-full max-w-2xl"
@@ -206,7 +40,7 @@ const TodoModal = () => {
             <Button
               className="rounded-xl py-2 px-4 bg-secondary text-white hover:bg-primary transition-colors"
               type="button"
-              onClick={onClose}
+              onClick={actions.onClose}
             >
               모달 닫기
             </Button>
@@ -214,7 +48,7 @@ const TodoModal = () => {
               <Button
                 className="rounded-xl py-2 px-4 bg-secondary text-white hover:bg-primary transition-colors"
                 type="button"
-                onClick={updateTodoInput}
+                onClick={actions.updateTodoInput}
               >
                 적용
               </Button>
@@ -223,7 +57,7 @@ const TodoModal = () => {
                 className="rounded-xl py-2 px-4 bg-secondary text-white hover:bg-primary transition-colors"
                 type="button"
                 onClick={() => {
-                  handleEditTodo(todo)
+                  actions.handleEditTodo(todo)
                 }}
               >
                 수정
@@ -232,7 +66,7 @@ const TodoModal = () => {
             <Button
               className="rounded-xl py-2 px-4 bg-danger text-white hover:opacity-80 transition-opacity"
               type="button"
-              onClick={() => setShowDeleteConfirm(true)}
+              onClick={() => actions.setShowDeleteConfirm(true)}
             >
               삭제
             </Button>
@@ -265,9 +99,9 @@ const TodoModal = () => {
                 <span className="text-gray-800">
                   {todo.memo_id &&
                   todo.memo_id.trim() !== '' &&
-                  todoMemo &&
-                  todoMemo.memo
-                    ? todoMemo.memo.content
+                  state.todoMemo &&
+                  state.todoMemo.memo
+                    ? state.todoMemo.memo.content
                     : '없음'}
                 </span>
               </div>
@@ -283,8 +117,8 @@ const TodoModal = () => {
                 <input
                   className="h-10 px-3 w-full rounded-xl text-gray-800 border border-gray-300 focus:outline-none focus:border-secondary focus:ring-2 focus:ring-secondary/20 transition-all font-nanumgothic_regular"
                   type="text"
-                  value={newTask}
-                  onChange={(e) => setNewTask(e.target.value)}
+                  value={state.newTask}
+                  onChange={(e) => actions.setNewTask(e.target.value)}
                 />
               </label>
               <div className="flex flex-wrap gap-x-4 gap-y-2 mb-4">
@@ -292,18 +126,18 @@ const TodoModal = () => {
                   <div className="font-nanumgothic_regular">중요도</div>
                   <input
                     type="checkbox"
-                    checked={newImportant}
+                    checked={state.newImportant}
                     className="size-6"
-                    onChange={(e) => setNewImportant(e.target.checked)}
+                    onChange={(e) => actions.setNewImportant(e.target.checked)}
                   />
                 </label>
                 <label className="inline-flex items-center whitespace-nowrap gap-2">
                   <div className="font-nanumgothic_regular">완료</div>
                   <input
                     type="checkbox"
-                    checked={newCompleted}
+                    checked={state.newCompleted}
                     className="size-6"
-                    onChange={(e) => setNewCompleted(e.target.checked)}
+                    onChange={(e) => actions.setNewCompleted(e.target.checked)}
                   />
                 </label>
               </div>
@@ -319,13 +153,13 @@ const TodoModal = () => {
                   <input
                     className="h-9 px-2 rounded-xl text-gray-800 border border-gray-300 focus:outline-none focus:border-secondary focus:ring-2 focus:ring-secondary/20 transition-all font-nanumgothic_regular"
                     type="date"
-                    value={newDate || ''}
-                    onChange={(e) => setNewDate(e.target.value || null)}
+                    value={state.newDate || ''}
+                    onChange={(e) => actions.setNewDate(e.target.value || null)}
                   />
                 </label>
                 <Button
                   type="button"
-                  onClick={handleClearDate}
+                  onClick={actions.handleClearDate}
                   className="py-2 px-4 rounded-xl bg-secondary text-white hover:bg-primary transition-colors"
                 >
                   날짜 미정
@@ -336,15 +170,15 @@ const TodoModal = () => {
                   연결 메모 추가/수정
                   <input
                     type="checkbox"
-                    checked={newConnect}
+                    checked={state.newConnect}
                     className="self-center size-6"
-                    onChange={(e) => setNewConnect(e.target.checked)}
+                    onChange={(e) => actions.setNewConnect(e.target.checked)}
                   />
                 </label>
                 <Button
                   type="button"
                   className="rounded-xl py-2 px-4 bg-secondary text-white hover:bg-primary transition-colors"
-                  onClick={NewConnectNull}
+                  onClick={actions.NewConnectNull}
                 >
                   메모 연결 초기화
                 </Button>
@@ -353,45 +187,45 @@ const TodoModal = () => {
                 <div>
                   {todo.memo_id && todo.memo_id.trim() !== '' && (
                     <div>
-                      {todoMemo && todoMemo.memo && (
+                      {state.todoMemo && state.todoMemo.memo && (
                         <div className="mb-4">
-                          기존 Memo의 Content : {todoMemo.memo.content}
+                          기존 Memo의 Content : {state.todoMemo.memo.content}
                         </div>
                       )}
                     </div>
                   )}
                 </div>
               </label>
-              {newConnect && (
+              {state.newConnect && (
                 <div>
                   <div className="flex items-center mb-4">
                     <div>새로 연동할 메모를 선택하세요</div>
                   </div>
-                  {newConnectMemoContent &&
-                    newConnectMemoContent.trim() !== '' && (
+                  {state.newConnectMemoContent &&
+                    state.newConnectMemoContent.trim() !== '' && (
                       <div className="mb-4 text-secondary bg-secondary/5 border border-secondary/20 rounded-lg px-3 py-2">
                         🔗 새로운 메모:{' '}
                         <span className="font-nanumgothic_bold">
-                          {newConnectMemoContent}
+                          {state.newConnectMemoContent}
                         </span>
                       </div>
                     )}
-                  {(!newConnectMemoContent ||
-                    newConnectMemoContent.trim() === '') && (
+                  {(!state.newConnectMemoContent ||
+                    state.newConnectMemoContent.trim() === '') && (
                     <div className="mb-4 font-nanumgothic_regular text-gray-500">
                       새로운 메모 :{' '}
-                      {isMemoNull ? <span>없음</span> : <span>❔</span>}
+                      {state.isMemoNull ? <span>없음</span> : <span>❔</span>}
                     </div>
                   )}
                 </div>
               )}
             </div>
           )}
-          {editTodo && newConnect && (
+          {editTodo && state.newConnect && (
             <div className="grid grid-cols-[repeat(auto-fit,minmax(9rem,1fr))] gap-2 sm:gap-4 justify-items-center">
-              {connectMemos.map((memo: Memo) => (
+              {state.connectMemos.map((memo: Memo) => (
                 <ConnectMemoBox
-                  memoFetch={() => MemoIDContent(memo.id, memo.content)}
+                  memoFetch={() => actions.MemoIDContent(memo.id, memo.content)}
                   key={memo.id}
                   memo={memo}
                 />
@@ -399,7 +233,7 @@ const TodoModal = () => {
             </div>
           )}
         </div>
-        {showDeleteConfirm && (
+        {state.showDeleteConfirm && (
           <div className="absolute inset-0 z-10 flex items-center justify-center rounded-2xl bg-black/40 backdrop-blur-sm">
             <div className="animate-fade-in-scale bg-white rounded-2xl p-8 shadow-xl flex flex-col items-center gap-3 text-center">
               <p className="font-nanumgothic_bold text-primary text-ui-sm">
@@ -412,14 +246,14 @@ const TodoModal = () => {
                 <Button
                   className="rounded-xl py-2 px-4 bg-gray-200 text-gray-700 hover:bg-gray-300 transition-colors"
                   type="button"
-                  onClick={() => setShowDeleteConfirm(false)}
+                  onClick={() => actions.setShowDeleteConfirm(false)}
                 >
                   취소
                 </Button>
                 <Button
                   className="rounded-xl py-2 px-4 bg-danger text-white hover:opacity-80 transition-opacity"
                   type="button"
-                  onClick={() => handleDeleteTodo(todo.id)}
+                  onClick={() => actions.handleDeleteTodo(todo.id)}
                 >
                   삭제 확인
                 </Button>
